Teams App to v0.8.2.4, Working on GUI, logic.
-Working on GUI!!! (finally)
-Working on logic (file team posts, team posts, function return values, and where files and data are stored).
-Added header, updated JScript.

+++ Applications/Teams/Teams.php
@@ -4,7 +4,7 @@
 /*//
 HRCLOUD2-PLUGIN-START
 App Name: Teams
-App Version: 0.8.2.3  (5-13-2017 00:00)
+App Version: v0.8.2.4 (5-13-2017 00:00)
 App License: GPLv3
 App Author: zelon88
 App Description: A simple HRCloud2 App for communicating with team-mates.
@@ -16,6 +16,7 @@ HRCLOUD2-PLUGIN-END
 <script src="_SCRIPTS/sorttable.js"></script>
 <script src="_SCRIPTS/clearField1.js"></script>
 <script src="_SCRIPTS/common.js"></script>
+<link rel='stylesheet' type='text/css' href='_SCRIPTS/style.php' />
 <?php
 
 // / The follwoing code checks if the teamsCore.php file exists and 
@@ -25,6 +26,9 @@ if (!file_exists('/var/www/html/HRProprietary/HRCloud2/Applications/Teams/teamsC
   die (); }
 else {
   require_once ('/var/www/html/HRProprietary/HRCloud2/Applications/Teams/teamsCore.php'); }
+
+if ($headerDivNeeded == 'true') {
+include($headerFile); }
 
 // / The following code represents the graphical user-interface (GUI).
 if ($teamsHeaderDivNeeded == 'true') {
@@ -39,7 +43,7 @@ if ($teamsGreetingDivNeeded == 'true') {
   echo ('<div align="center" style="width:90%; float:center; cursor:pointer;">'.$emptyTeamsECHO.'</div></div>'); } }
 
 if ($newTeamDivNeeded == 'true') {
-  echo ('<div id="newTeamsDiv" name="newTeamsDiv" align="center" style="width:65%; float:center; display:none; border: 1px solid '.$color.'; border-radius: 6px;">
+  echo ('<div align=\'center\'><div id="newTeamsDiv" name="newTeamsDiv" align="center" style="width:65%; display:none; border: 1px solid '.$color.'; border-radius: 6px;">
     <img title="Close \'New Teams\'" alt="Close \'New Teams\'" id="xNewTeams1" name="xNewTeams1" style="float:right; display:none;" onclick="toggle_visibility(\'newTeamsDiv\'); toggle_visibility(\'xNewTeams1\');" src="_RESOURCES/x.png">
     <form method="post" action="Teams.php" type="multipart/form-data"><h4>New Team</h4>');
   echo nl2br('<input type="text" id="newTeam" name="newTeam" value="'.$newTeamNameEcho.'" onclick="Clear1();">'."\n");
@@ -48,7 +52,19 @@ if ($newTeamDivNeeded == 'true') {
     <option value="0">Public</option>
     <option value="1">Private</option>
     </select>'."\n");
-  echo nl2br('<input type="submit" id="newTeamButton" name="newTeamButton" value="New Team"></form></div>'."\n"); }
+  echo nl2br('<input type="submit" id="newTeamButton" name="newTeamButton" value="New Team"></form></div></div>'."\n"); }
+
+if ($teamsDivNeeded == 'true') { 
+
+}
+
+if ($friendsDivNeeded == 'true') { 
+
+}
+
+if ($filesDivNeeded == 'true') { 
+
+}
 
 if (isset($friendToAdd) && $friendToAdd !== '') {
   addFriend($friendToAdd); }
@@ -62,7 +78,7 @@ if (isset($newTeamName) && $newTeamName !== '') {
 if (isset($teamToEdit) && $teamToEdit !== '') {
   editTeam($teamToEdit); }
 
-if (isset($teamToDelete)) {
+if (isset($teamToDelete) && $teamToDelete !== '') {
   deleteTeam($teamToDelete); }
 
 if (isset($adminAddUserToTeam) && isset($adminTeamToAdd)) {
@@ -71,14 +87,14 @@ if (isset($adminAddUserToTeam) && isset($adminTeamToAdd)) {
 if (isset($adminRemoveUser) && isset($adminTeamToRemove)) {
   adminRemoveUser($adminRemoveUser, $adminTeamToRemove); }
 
-if (isset($newSubTeam) && $newSubTeam !== '' && isset($teamToJoin)) {
-  createNewSubTeam($newSubTeam, $subTeamUsers); }
-
-if (isset($subTeamToJoin) && $subTeamToJoin !== '' && isset($teamToJoin)) {
-  joinSubTeam($teamToJoin, $subTeamToJoin); }
-
 if (isset($teamToJoin) && $teamToJoin !== '') {
   joinTeam($teamToJoin); }
+
+if (isset($newSubTeam) && $newSubTeam !== '' && isset($teamToJoin) && $teamToJoin !== '') {
+  createNewSubTeam($newSubTeam, $subTeamUsers); }
+
+if (isset($subTeamToJoin) && $subTeamToJoin !== '' && isset($teamToJoin) && $teamToJoin !== '') {
+  joinSubTeam($teamToJoin, $subTeamToJoin); }
 
 
 
@@ -87,36 +103,4 @@ if ($myTeamsDivNeeded == 'true') {
 
 if ($teamsDivNeeded == 'true') {
   getPublicTeams($teamsList); } 
-?>
-
-
-<?php
-/*
-</div><div id="TeamsList" name="TeamsList" align="center"><strong>Teams</strong><hr /></div>
-<div align="center">
-<table class="sortable">
-<thead><tr>
-<th>Team</th>
-<th>Edit</th>
-<th>Delete</th>
-<th>Last Modified</th>
-</tr></thead><tbody>
- <?php 
-$notesList2 = scandir($NotesDir); 
-$noteCounter = 0;
-foreach ($notesList2 as $note) {
-  if ($note == '.' or $note == '..' or strpos($note, '.txt') == 'false' 
-    or $note == '' or $note == '.txt') continue; 
-  $noteCounter++;
-  $noteFile = $NotesDir.$note; 
-  $noteEcho = str_replace('.txt', '', $note);
-  $noteTime = date("F d Y H:i:s.",filemtime($noteFile));
-  echo nl2br ('<tr><td><strong>'.$noteCounter.'. </strong><a href="Notes.php?editNote='.$noteEcho.'">'.$noteEcho.'</a></td>');
-  echo nl2br('<td><a href="Notes.php?editNote='.$noteEcho.'"><img id="edit'.$noteCounter.'" name="'.$note.'" src="'.$URL.'/HRProprietary/HRCloud2/Resources/edit.png"></a></td>');
-  echo nl2br('<td><a href="Notes.php?deleteNote='.$noteEcho.'"><img id="delete'.$noteCounter.'" name="'.$note.'" src="'.$URL.'/HRProprietary/HRCloud2/Resources/deletesmall.png"></a></td>');
-  echo nl2br('<td><a><i>'.$noteTime.'</i></a></td></tr>'); } ?>
-<tbody>
-</table>
-</div>
-*/
